Refactored post preview modal use
Slimmed down page object to only what's currently used; can add other parts later
